Now does not break on AI turns, will play hundreds of runes until program kill. Bias to keep board clear in both players.

+++ src/runes/modify_hero_health.rs
@@ -8,13 +8,13 @@ use hlua;
 #[derive(RustcDecodable, RustcEncodable, Clone, Debug)]
 pub struct ModifyHeroHealth {
     target_uid: UID,
-    amount: u8,
+    amount: i32,
 }
 
 implement_for_lua!(ModifyHeroHealth, |mut _metatable| {});
 
 impl ModifyHeroHealth {
-    pub fn new(target_uid: UID, amount: u8) -> ModifyHeroHealth {
+    pub fn new(target_uid: UID, amount: i32) -> ModifyHeroHealth {
         ModifyHeroHealth {
             target_uid: target_uid,
             amount: amount,
